chore(benches): fix warnings and expose server crate as a library
Rework benchmark code to compile cleanly against the current crate
shape and APIs, and promote `rs-broker-server` to a library crate so
its gRPC service can be reused from benches.

- Add `src/lib.rs` plus `[lib]` and `[[bench]]` entries in
  `rs-broker-server/Cargo.toml`; enable the `async_tokio` feature
  for criterion
- Replace ad-hoc `setup_test_db` / string-URL pool creation with
  `DatabaseConfig` + `create_pool` across the server benches
- Switch proto request types to the generated tonic service trait
  inputs and update optional string fields to `String::new()`
- Implement the new `create_batch` method in mock `OutboxRepository`
  impls used by core/kafka benches
- Suppress unused-variable warnings (`let _ = rt.block_on(...)`,
  `_repo`) and apply idiomatic patterns (`or_default`,
  `sort_by_key(Reverse)`, `100_f64`)
- Drop the unused `BaseProducer` import and unused `RetryStrategy`
  import in the kafka/outbox benches
- Split `RUST_LOG` into a dedicated `x-rust-env` anchor and add
  `TZ: Asia/Bangkok` to the shared `x-common-env` in `compose.yml`

+++ crates/rs-broker-kafka/benches/e2e.rs
@@ -118,7 +118,7 @@ fn bench_e2e_throughput(c: &mut Criterion) {
             }
 
             let elapsed = start.elapsed();
-            let throughput = 100 as f64 / elapsed.as_secs_f64();
+            let throughput = 100_f64 / elapsed.as_secs_f64();
             black_box(throughput)
         })
     });
